Move out general-purpose functions from runVisualDiffs
This work was prepared by @lelandrichardson as part of PR #179. I'm
cherry-picking them here to simplify work going forward.

These methods are all general-purpose, and should be possible to reuse
over multiple targets.
